requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
huggingface-hub>=0.19.0
python-dateutil>=2.8.2
jinja2>=3.1.2
click>=8.1.0
anthropic>=0.39.0
cohere>=5.11.0
groq>=0.11.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
